% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bg.R
\name{run_bg}
\alias{run_bg}
\alias{run_view}
\alias{run_bg_view}
\title{Use background jobs interactively}
\usage{
run_bg(cntx = NULL, view = FALSE)

run_view(cntx = NULL)

run_bg_view(cntx = NULL)
}
\arguments{
\item{cntx}{a string param containing code to launch (mostly for debug purpose)}

\item{view}{a boolean param available only for \code{run_bg()}}
}
\value{
\code{run_bg()} and \code{run_bg_view} returns objects from background environment to global one
\code{run_bg_view} and \code{run_view} launch \code{View()} function to show result
}
\description{
Set of function to make interactive work faster
}
\details{
Group of functions Details paragraph.
}
\section{Functions}{
\itemize{
\item \code{run_bg()}: Runs the selected code as a background job importing objects and packages from original session

\item \code{run_view()}: Runs the selected code and then performs View

\item \code{run_bg_view()}: Runs the selected code as a background job importing objects and packages from original session and finally views a result

}}
